+++ docs/docs/changelog.md
@@ -38,6 +38,7 @@ SOFTWARE.
 -----
 Released: Under development
 
+* Bugfix: Added default URL to use if `site_url` is not defined in mkdocs.yml. Fix for issue [#15](https://github.com/iSOLveIT/mkdocs-pdf-generate/issues/15).
 * Improvement: Added the `build_txt` local option that allows users to build TXT file containing the Table of Contents
 * Minor improvements: Rewrote logging messages.